refactor(forms): extract form validation logic to reusable hook
Extract repeated form validity logic from CreateOrUpdateModal and
AddEntityToCollectionsForm into a shared useFormValidation hook.
