+++ apps/frontend/app/lib/shared/hooks.ts
@@ -13,6 +13,7 @@ import {
 import { useMutation, useQuery } from "@tanstack/react-query";
 import Cookies from "js-cookie";
 import type { FormEvent } from "react";
+import { useCallback, useRef, useState } from "react";
 import { useNavigate } from "react-router";
 import {
 	useRevalidator,
@@ -330,4 +331,17 @@ export const useRemoveEntitiesFromCollectionMutation = () => {
 	});
 
 	return mutation;
+};
+
+export const useFormValidation = () => {
+	const formRef = useRef<HTMLFormElement>(null);
+	const [isFormValid, setIsFormValid] = useState(true);
+
+	const checkFormValidity = useCallback(() => {
+		if (formRef.current) {
+			setIsFormValid(formRef.current.checkValidity());
+		}
+	}, []);
+
+	return { formRef, isFormValid, checkFormValidity };
 };
